examples/python/snippets/get-started/stock_agent/requirements.txt
google-genai-agents
yfinance